Update Issue.all_addressed
There is an issue with my "not in array" tasks query for non-closed tasks.
In dev, I get the right result if I exclude tasks that have a NULL
closed.  In test, I need to exclude tasks that have a false closed.  For
some reason, I can't get them to both work together.  Added workaround
based on the env for now.  Will add counter caches for the tasks count
that will make theses queries simpler.