Add github_user_id column to issues
GitHub webhook doesn't include the user email so can't add the real user,
so using a common "bot" user for now.

Planning on sending requests to their API so that I can add a comment to
the imported issue. Maybe I can find out the email in another request
and invite the user to join the app. Or just leave a comment on github
issues that lets them join the app. If the user joins, I can provide a
way to connect the imported issues to their account. Probably need oauth
to connect app user to github user id.

In the meantime, saving the github user id to the issue